move patch.ts to Utils.ts
