+++ tools/src/Utils.ts
@@ -140,3 +140,29 @@ export function arrayize<T>(value: T | T[]): T[] {
   }
   return value != null ? [value] : [];
 }
+
+/**
+ * Execute `patch` command for given patch content
+ */
+export async function applyPatchAsync(options: {
+  patchContent: string;
+  cwd: string;
+  reverse?: boolean;
+  stripPrefixNum?: number;
+}) {
+  const args: string[] = [];
+  if (options.stripPrefixNum != null) {
+    // -pN passing to the `patch` command for striping slashed prefixes
+    args.push(`-p${options.stripPrefixNum}`);
+  }
+  if (options.reverse) {
+    args.push('-R');
+  }
+
+  const procPromise = spawnAsync('patch', args, {
+    cwd: options.cwd,
+  });
+  procPromise.child.stdin?.write(options.patchContent);
+  procPromise.child.stdin?.end();
+  await procPromise;
+}
